+++ CHANGELOG.md
@@ -28,6 +28,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
   Replaces any language directive in the YAML file.
 
 ### Changed
+- Default of library name from *default_library* to *library*.
 - Renamed option *C_header_helper_template* to *C_header_utility_template*.
   Renamed option *PY_helper_filename_template* to *PY_utililty_filename_template*.
   This is to avoid confusion with helper functions which have file static scope.